split MINIT for WeekInfo
