+++ src/php/ecma_intl_arginfo.h
@@ -1,5 +1,5 @@
 /* This is a generated file, edit the .stub.php file instead.
- * Stub hash: 7ddfa21a419a4884073bf5764537976a096ec5d3 */
+ * Stub hash: 072a04a13d5217758e021d976cd7ca954937a184 */
 
 ZEND_BEGIN_ARG_WITH_RETURN_TYPE_INFO_EX(arginfo_class_Ecma_Intl_getCanonicalLocales, 0, 1, IS_ARRAY, 0)
 	ZEND_ARG_OBJ_TYPE_MASK(0, locales, Traversable|Stringable, MAY_BE_ARRAY|MAY_BE_STRING|MAY_BE_NULL, NULL)
@@ -83,14 +83,6 @@ ZEND_END_ARG_INFO()
 
 #define arginfo_class_Ecma_Intl_Locale_TextInfo_jsonSerialize arginfo_class_Ecma_Intl_Locale_jsonSerialize
 
-ZEND_BEGIN_ARG_INFO_EX(arginfo_class_Ecma_Intl_Locale_WeekInfo___construct, 0, 0, 3)
-	ZEND_ARG_OBJ_INFO(0, firstDay, Ecma\\Intl\\Locale\\WeekDay, 0)
-	ZEND_ARG_TYPE_INFO(0, weekend, IS_ARRAY, 0)
-	ZEND_ARG_TYPE_INFO(0, minimalDays, IS_LONG, 0)
-ZEND_END_ARG_INFO()
-
-#define arginfo_class_Ecma_Intl_Locale_WeekInfo_jsonSerialize arginfo_class_Ecma_Intl_Locale_jsonSerialize
-
 
 ZEND_METHOD(Ecma_Intl, getCanonicalLocales);
 ZEND_METHOD(Ecma_Intl, supportedValuesOf);
@@ -117,8 +109,6 @@ ZEND_METHOD(Ecma_Intl_Locale_Options, valid);
 ZEND_METHOD(Ecma_Intl_Locale_Options, rewind);
 ZEND_METHOD(Ecma_Intl_Locale_TextInfo, __construct);
 ZEND_METHOD(Ecma_Intl_Locale_TextInfo, jsonSerialize);
-ZEND_METHOD(Ecma_Intl_Locale_WeekInfo, __construct);
-ZEND_METHOD(Ecma_Intl_Locale_WeekInfo, jsonSerialize);
 
 
 static const zend_function_entry class_Ecma_Intl_methods[] = {
@@ -178,13 +168,6 @@ static const zend_function_entry class_Ecma_Intl_Locale_TextInfo_methods[] = {
 
 
 static const zend_function_entry class_Ecma_Intl_Locale_WeekDay_methods[] = {
-	ZEND_FE_END
-};
-
-
-static const zend_function_entry class_Ecma_Intl_Locale_WeekInfo_methods[] = {
-	ZEND_ME(Ecma_Intl_Locale_WeekInfo, __construct, arginfo_class_Ecma_Intl_Locale_WeekInfo___construct, ZEND_ACC_PUBLIC)
-	ZEND_ME(Ecma_Intl_Locale_WeekInfo, jsonSerialize, arginfo_class_Ecma_Intl_Locale_WeekInfo_jsonSerialize, ZEND_ACC_PUBLIC)
 	ZEND_FE_END
 };
 
@@ -505,37 +488,6 @@ static zend_class_entry *register_class_Ecma_Intl_Locale_WeekDay(void)
 	zval enum_case_Sunday_value;
 	ZVAL_LONG(&enum_case_Sunday_value, 7);
 	zend_enum_add_case_cstr(class_entry, "Sunday", &enum_case_Sunday_value);
-
-	return class_entry;
-}
-
-static zend_class_entry *register_class_Ecma_Intl_Locale_WeekInfo(zend_class_entry *class_entry_JsonSerializable)
-{
-	zend_class_entry ce, *class_entry;
-
-	INIT_NS_CLASS_ENTRY(ce, "Ecma\\Intl\\Locale", "WeekInfo", class_Ecma_Intl_Locale_WeekInfo_methods);
-	class_entry = zend_register_internal_class_ex(&ce, NULL);
-	class_entry->ce_flags |= ZEND_ACC_READONLY_CLASS;
-	zend_class_implements(class_entry, 1, class_entry_JsonSerializable);
-
-	zend_string *property_firstDay_class_Ecma_Intl_Locale_WeekDay = zend_string_init("Ecma\\Intl\\Locale\\WeekDay", sizeof("Ecma\\Intl\\Locale\\WeekDay")-1, 1);
-	zval property_firstDay_default_value;
-	ZVAL_UNDEF(&property_firstDay_default_value);
-	zend_string *property_firstDay_name = zend_string_init("firstDay", sizeof("firstDay") - 1, 1);
-	zend_declare_typed_property(class_entry, property_firstDay_name, &property_firstDay_default_value, ZEND_ACC_PUBLIC|ZEND_ACC_READONLY, NULL, (zend_type) ZEND_TYPE_INIT_CLASS(property_firstDay_class_Ecma_Intl_Locale_WeekDay, 0, 0));
-	zend_string_release(property_firstDay_name);
-
-	zval property_minimalDays_default_value;
-	ZVAL_UNDEF(&property_minimalDays_default_value);
-	zend_string *property_minimalDays_name = zend_string_init("minimalDays", sizeof("minimalDays") - 1, 1);
-	zend_declare_typed_property(class_entry, property_minimalDays_name, &property_minimalDays_default_value, ZEND_ACC_PUBLIC|ZEND_ACC_READONLY, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_LONG));
-	zend_string_release(property_minimalDays_name);
-
-	zval property_weekend_default_value;
-	ZVAL_UNDEF(&property_weekend_default_value);
-	zend_string *property_weekend_name = zend_string_init("weekend", sizeof("weekend") - 1, 1);
-	zend_declare_typed_property(class_entry, property_weekend_name, &property_weekend_default_value, ZEND_ACC_PUBLIC|ZEND_ACC_READONLY, NULL, (zend_type) ZEND_TYPE_INIT_MASK(MAY_BE_ARRAY));
-	zend_string_release(property_weekend_name);
 
 	return class_entry;
 }
